# tests: no default interactive sync for native