--- a/tests/Makefile.tests_common
+++ b/tests/Makefile.tests_common
@@ -1,17 +1,20 @@
 APPLICATION ?= tests_$(notdir $(patsubst %/,%,$(CURDIR)))
-
-ifneq (,$(wildcard $(CURDIR)/tests*/.))
-  # add interactive test configuration when testing Kconfig
-  DEFAULT_MODULE += test_utils_interactive_sync
-  # add stack metric printing configuration when testing Kconfig
-  DEFAULT_MODULE += test_utils_print_stack_usage
-endif
-
-# terminate native when the test is complete
-CFLAGS += -DNATIVE_AUTO_EXIT=1
-
 BOARD ?= native
 RIOTBASE ?= $(CURDIR)/../..
 QUIET ?= 1
 # DEVELHELP enabled by default for all tests, set 0 to disable
 DEVELHELP ?= 1
+
+ifneq (,$(wildcard $(CURDIR)/tests*/.))
+  # add interactive test configuration when automated testing is available,
+  # except for `native` and `native64` where we control when the RIOT app is
+  # started
+  ifeq (,$(filter native native64,$(BOARD)))
+    DEFAULT_MODULE += test_utils_interactive_sync
+  endif
+  # print stack usage by default when automated testing is available
+  DEFAULT_MODULE += test_utils_print_stack_usage
+endif
+
+# terminate native when the test is complete
+CFLAGS += -DNATIVE_AUTO_EXIT=1