@maribu 2024-11-08
> `make NOSYNC=1 test` will surprisingly not work

That is expected: `make test` will run whatever binary is in place. If the `make test` is run with `NOSYNC=1` but `make all` is run with `NOSYNC=0`, the test will expect no interactive sync, but the binary will provide it.

> and for all cases where the Python tests do more than just grepping for OK (xx tests), this will need to be done manually.

True, but native does actually pass through the return value of the `main` thread.

In my experience tests that do rely heavily on python scripting are both slow and fragile, especially since stdio can be unreliable (e.g. bit errors on UART is a thing). For these reasons, the simple test style where the python only checks for some `TEST PASSED` or `TEST SUCCEEDED` in the output is becoming more and more popular.

> This adds another obscure environment variable that is not documented anywhere except in the Makefile

True. Maybe we should just disable interactive sync for `native` and `native64` unconditionally instead? I think there is no reason to sync with native anyway.