@mguetschow 2024-11-08
> > `make NOSYNC=1 test` will surprisingly not work
> 
> That is expected: `make test` will run whatever binary is in place. If the `make test` is run with `NOSYNC=1` but `make all` is run with `NOSYNC=0`, the test will expect no interactive sync, but the binary will provide it.

Ah, that's not what I meant, I thought the Python testrunner would always expect the interactive sync. According to the docs in `tests/README.md` it should actually adapt automatically if the module is disabled. So please disregard this comment, and also the follow up of it (the manual vs. Python grepping one).

> > This adds another obscure environment variable that is not documented anywhere except in the Makefile
> 
> True. Maybe we should just disable interactive sync for `native` and `native64` unconditionally instead? I think there is no reason to sync with native anyway.

Sounds like a plan. Maybe check if you find a nice place to document this, either for the native board or in the above-mentioned tests README.