@maribu 2024-11-08
The tests on `native` and `native64` are run no matter what, but let's also test for regressions.

I think it is fine to have the `CI: run tests` set only for the quick run - after all, this change would effect all tests and boards, if this introduces issues. I'll drop that label when this passes.